폴더 재구성 및 리프레시 토큰 재발급, 이메일 중복 확인 기능 추가
