+++ src/main/java/org/example/hungryback/dto/request/account/PatchPasswordRequestDto.java
@@ -1,7 +1,8 @@
-package org.example.hungryback.dto.request.user;
+package org.example.hungryback.dto.request.account;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -9,12 +10,14 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class PatchProfileImgRequestDto {
+public class PatchPasswordRequestDto {
 
     @Email
     @NotBlank
     private String userEmail;
 
-    private String userProfileImg;
+    @NotBlank
+    @Pattern(regexp="^(?=.*[a-zA-Z])(?=.*[0-9])[a-zA-Z0-9]{8,13}$")
+    private String userPassword;
 
 }
