+++ src/main/java/org/example/hungryback/dto/request/account/PatchProfileImgRequestDto.java
@@ -1,8 +1,7 @@
-package org.example.hungryback.dto.request.user;
+package org.example.hungryback.dto.request.account;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -10,14 +9,12 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class PatchPasswordRequestDto {
+public class PatchProfileImgRequestDto {
 
     @Email
     @NotBlank
     private String userEmail;
 
-    @NotBlank
-    @Pattern(regexp="^(?=.*[a-zA-Z])(?=.*[0-9])[a-zA-Z0-9]{8,13}$")
-    private String userPassword;
+    private String userProfileImg;
 
 }
