+++ src/main/java/org/example/hungryback/dto/request/auth/CheckCertificationRequestDto.java
@@ -1,24 +1,20 @@
-package org.example.hungryback.dto.request.user;
+package org.example.hungryback.dto.request.auth;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Getter
 @Setter
-@NoArgsConstructor
-public class SignUpSocialRequestDto {
-    @NotBlank
-    private String userName;
+public class CheckCertificationRequestDto {
 
     @NotBlank
     @Pattern(regexp="^(01[016789]{1})[0-9]{7,8}$")
     private String userTel;
 
-    private String userProfileImg;
-
     @NotBlank
-    private String userNickname;
+    @Pattern(regexp="^[0-9]{6}$")
+    private String certificationNumber;
+
 }
