+++ src/main/java/org/example/hungryback/dto/request/auth/CheckEmailRequestDto.java
@@ -1,0 +1,14 @@
+package org.example.hungryback.dto.request.auth;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class CheckEmailRequestDto {
+    @Email
+    @NotBlank
+    private String userEmail;
+}
