--- src/main/java/org/example/hungryback/dto/request/certification/CheckCertificationRequestDto.java
@@ -1,20 +1,0 @@
-package org.example.hungryback.dto.request.certification;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CheckCertificationRequestDto {
-
-    @NotBlank
-    @Pattern(regexp="^(01[016789]{1})[0-9]{7,8}$")
-    private String userTel;
-
-    @NotBlank
-    @Pattern(regexp="^[0-9]{6}$")
-    private String certificationNumber;
-
-}
