+++ src/main/java/org/example/hungryback/dto/response/auth/CheckEmailResponseDto.java
@@ -1,0 +1,25 @@
+package org.example.hungryback.dto.response.auth;
+
+import lombok.Getter;
+import org.example.hungryback.common.ResponseCode;
+import org.example.hungryback.common.ResponseMessage;
+import org.example.hungryback.dto.ResponseDto;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+@Getter
+public class CheckEmailResponseDto extends ResponseDto {
+    private CheckEmailResponseDto() {
+        super();
+    }
+
+    public static ResponseEntity<CheckEmailResponseDto> success () {
+        CheckEmailResponseDto responseBody = new CheckEmailResponseDto();
+        return ResponseEntity.status(HttpStatus.OK).body(responseBody);
+    }
+
+    public static ResponseEntity<ResponseDto> duplicateEmail () {
+        ResponseDto responseBody = new ResponseDto(ResponseCode.DUPLICATE_EMAIL, ResponseMessage.DUPLICATE_EMAIL);
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(responseBody);
+    }
+}
