+++ src/main/java/org/example/hungryback/entity/CustomOAuth2User.java
@@ -1,6 +1,7 @@
 package org.example.hungryback.entity;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.user.OAuth2User;
@@ -12,6 +13,8 @@ import java.util.Map;
 @AllArgsConstructor
 public class CustomOAuth2User implements OAuth2User {
     private String userEmail;
+    @Getter
+    private String userType;
 
     @Override
     public Map<String, Object> getAttributes() {
@@ -27,4 +30,5 @@ public class CustomOAuth2User implements OAuth2User {
     public String getName() {
         return this.userEmail;
     }
+
 }
