+++ src/main/java/org/example/hungryback/entity/RefreshTokenEntity.java
@@ -5,15 +5,14 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.redis.core.RedisHash;
 import org.springframework.data.redis.core.index.Indexed;
 
-import java.io.Serializable;
-
 @Getter
 @NoArgsConstructor (access = AccessLevel.PROTECTED)
-@RedisHash(value = "RefreshToken", timeToLive = 60)
+@RedisHash(value = "RefreshToken", timeToLive = 60 * 60 * 24 * 14)
 public class RefreshTokenEntity {
     @Id
-    private String refreshToken;
     private String userEmail;
+    @Indexed
+    private String refreshToken;
 
     @Builder
     public RefreshTokenEntity(String refreshToken, String userEmail) {
