+++ src/main/java/org/example/hungryback/service/implement/AuthServiceImplement.java
@@ -2,16 +2,19 @@ package org.example.hungryback.service.implement;
 
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.example.hungryback.dto.request.user.*;
-import org.example.hungryback.dto.response.user.*;
 import org.example.hungryback.dto.ResponseDto;
+import org.example.hungryback.dto.request.auth.*;
+import org.example.hungryback.dto.response.auth.*;
+import org.example.hungryback.entity.CertificationEntity;
 import org.example.hungryback.entity.RefreshTokenEntity;
-import org.example.hungryback.provider.JwtProvider;
-import org.example.hungryback.repository.RefreshTokenRepository;
 import org.example.hungryback.entity.UserEntity;
+import org.example.hungryback.provider.CertificationProvider;
+import org.example.hungryback.provider.JwtProvider;
 import org.example.hungryback.repository.CertificationRepository;
+import org.example.hungryback.repository.RefreshTokenRepository;
 import org.example.hungryback.repository.UserRepository;
-import org.example.hungryback.service.UserService;
+import org.example.hungryback.service.AuthService;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -19,19 +22,23 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class UserServiceImplement implements UserService {
+public class AuthServiceImplement implements AuthService {
+
+    @Value("${jwt-access-expiration-time}")
+    private int tokenExpiredMs;
 
     private final UserRepository userRepository;
     private final RefreshTokenRepository refreshTokenRepository;
     private final CertificationRepository certificationRepository;
 
     private final JwtProvider jwtProvider;
+    private final CertificationProvider certificationProvider;
 
     private PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
 
-    // 앱 회원가입
+    // 회원가입
     @Override
-    public ResponseEntity<? super SignUpResponseDto> signUpApp(SignUpAppRequestDto dto) {
+    public ResponseEntity<? super SignUpResponseDto> signUp(SignUpRequestDto dto) {
         try {
 
             String userEmail = dto.getUserEmail();
@@ -47,24 +54,6 @@ public class UserServiceImplement implements UserService {
             dto.setUserPassword(encodedPassword);
 
             UserEntity userEntity = new UserEntity(dto);
-            userRepository.save(userEntity);
-
-        } catch (Exception exception) {
-            exception.printStackTrace();
-            return ResponseDto.databaseError();
-        }
-        return SignUpResponseDto.success();
-    }
-
-    // 소셜 로그인 회원가입
-    @Override
-    public ResponseEntity<? super SignUpResponseDto> signUpSocial(SignUpSocialRequestDto dto, String userEmail) {
-        try {
-
-            UserEntity userEntity = userRepository.findByUserEmail(userEmail);
-            if(userEntity == null) return SignUpResponseDto.notExistUser();
-
-            userEntity.signUpUser(dto);
             userRepository.save(userEntity);
 
         } catch (Exception exception) {
@@ -103,89 +92,70 @@ public class UserServiceImplement implements UserService {
             ResponseDto.databaseError();
         }
 
-        return SignInResponseDto.success(token);
+        return SignInResponseDto.success(token, tokenExpiredMs);
     }
 
+    // 이메일 중복 체크
     @Override
-    public ResponseEntity<? super PatchNicknameResponseDto> patchNickname(PatchNicknameRequestDto dto) {
+    public ResponseEntity<? super CheckEmailResponseDto> checkEmail(CheckEmailRequestDto dto) {
         try {
 
             String userEmail = dto.getUserEmail();
-            String userNickname = dto.getUserNickname();
-
-            UserEntity userEntity = userRepository.findByUserEmail(userEmail);
-            if(userEntity == null) return PatchNicknameResponseDto.notExistUser();
-
-            userEntity.patchNickname(userNickname);
-
-            userRepository.save(userEntity);
+            boolean isExistEmail = userRepository.existsByUserEmail(userEmail);
+            if(isExistEmail) return CheckEmailResponseDto.duplicateEmail();
 
         } catch (Exception exception) {
             exception.printStackTrace();
             return ResponseDto.databaseError();
         }
-        return PatchNicknameResponseDto.success();
+        return CheckEmailResponseDto.success();
     }
 
+    // 인증번호 전송
     @Override
-    public ResponseEntity<? super PatchProfileImgResponseDto> patchProfileImg(PatchProfileImgRequestDto dto) {
-
+    public ResponseEntity<? super SendCertificationResponseDto> sendCertification(SendCertificationRequestDto dto) {
         try {
 
-            String userEmail = dto.getUserEmail();
-            String userProfileImg = dto.getUserProfileImg();
+            String userTel = dto.getUserTel();
+            String certificationNumber = certificationProvider.createRandomNumber();
 
-            UserEntity userEntity = userRepository.findByUserEmail(userEmail);
-            if(userEntity == null) return PatchProfileImgResponseDto.notExistUser();
+            boolean isSuccessed = certificationProvider.sendCertification(userTel, certificationNumber);
+            if(!isSuccessed) return SendCertificationResponseDto.messageSendFail();
 
-            userEntity.patchProfileImg(userProfileImg);
-
-            userRepository.save(userEntity);
+            CertificationEntity certificationEntity = new CertificationEntity(userTel, certificationNumber);
+            certificationRepository.save(certificationEntity);
 
         } catch (Exception exception) {
             exception.printStackTrace();
             return ResponseDto.databaseError();
         }
-        return PatchProfileImgResponseDto.success();
+        return SendCertificationResponseDto.success();
     }
 
+    // 인증번호 확인
     @Override
-    public ResponseEntity<? super PatchPasswordResponseDto> patchPassword(PatchPasswordRequestDto dto) {
+    public ResponseEntity<? super CheckCertificationResponseDto> checkCertification(CheckCertificationRequestDto dto) {
+
+        String userEmail = null;
 
         try {
 
-            String userEmail = dto.getUserEmail();
-            String userPassword = dto.getUserPassword();
-            String encodedPassword = passwordEncoder.encode(userPassword);
+            String userTel = dto.getUserTel();
+            String certificationNumber = dto.getCertificationNumber();
 
-            UserEntity userEntity = userRepository.findByUserEmail(userEmail);
-            if(userEntity == null) return PatchProfileImgResponseDto.notExistUser();
+            CertificationEntity certificationEntity = certificationRepository.findByUserTel(userTel);
+            if(certificationEntity == null || !certificationEntity.getCertificationNumber().equals(certificationNumber))
+                return CheckCertificationResponseDto.certificationFail();
 
-            userEntity.patchPassword(encodedPassword);
+            certificationRepository.delete(certificationEntity);
 
-            userRepository.save(userEntity);
+            UserEntity userEntity = userRepository.findByUserTel(userTel);
+            if(userEntity != null ) userEmail = userEntity.getUserEmail();
 
         } catch (Exception exception) {
             exception.printStackTrace();
             return ResponseDto.databaseError();
         }
-        return PatchPasswordResponseDto.success();
-    }
-
-    @Override
-    public ResponseEntity<? super GetUserResponseDto> getUser(String userEmail) {
-
-        UserEntity userEntity;
-
-        try {
-
-            userEntity = userRepository.findByUserEmail(userEmail);
-            if(userEntity == null) return GetUserResponseDto.notExistUser();
-
-        } catch (Exception exception) {
-            exception.printStackTrace();
-            return GetUserResponseDto.databaseError();
-        }
-        return GetUserResponseDto.success(userEntity);
+        return CheckCertificationResponseDto.success(userEmail);
     }
 }
