+++ src/main/resources/DDL.sql
@@ -1,4 +1,5 @@
 use hungry;
 
 select * from user;
+delete from user;
